requirements.txt
pandas
requests
seaborn
matplotlib
python-dotenv

